refactor(studio): decompose Canvas; grouped palette + unified presets UI
Canvas.tsx (4,396 lines, ~18 responsibilities) drops to 3,344: themes,
sacks, saved layouts, background, def-editor entry points and splitter
drags extract to canvas/use*.ts hooks in the established deps-object
style; the Settings/Background/Plugins/DesignerList sections become
lazy-loaded panel components. useEditorModel splits into pure
canvas/editorOps.ts (~50 op helpers) plus grouped sub-reducers — every
import site and all ten co-located test files unchanged.

The Inspector Add palette renders grouped by meta.category (Meters /
Clocks / Utility / per-plugin name; pure groupPalette helper, tested);
per-widget theme overrides collapse behind a summary with an "N set"
count, pre-opened when overrides exist; templates get preview jumps into
the designer and explicit standalone-copy vs clone-to-library labelling;
the title bar drops its permanently disabled Cancel.
